services/vision/requirements.txt
# Comunicação
nats-py

# Processamento de Imagem
numpy
opencv-python-headless
Pillow

# IA / OCR
easyocr
